API to DB tests step 1: authentication harness
This commit is the first step towards having API to DB tests running in the
ee2 test suite. As part of a noop test, it starts and stops the auth
service.

Tests were manually verified to pass when run inside the Dockerfile.

The noop test was manually verified to print out the root HTML page from
the auth server.