Move debug mode to job requirements
After a bit of cogitation it makes more sense there since it applies to
how the job is run. In any case, it's stored in the catalog along with all the
other params in the class, so it also makes sense practially.

Also made the client group regex have a 'no preference' default value so the
scheduler wrapper can decide on a sane default for the scheulder (for Condor
the default is to use a regex, for example, which to me isn't intuitive, and
I feel iffy about making the default True).