Fix bug processing catalog client groups in CSV format
Details: https://github.com/kbase/execution_engine2/issues/332

In short, the CSV inputs shown in the catalog UI are already split into
separate strings before being sent to the catalog UI, so the split is not
necesssary. Only accepting the first entry in the list causes everything
except the client group to be ignored.