Add fn(job_requirements) -> requirements type
Determine a requirements type based on job requirements. This will be used
to determine whether a user is allowed to run a job with non-default
requirements based on their ee2 admin status.

*Probably* for now anything other than standard will require write admin but
billing is split out as billing a job to another user seems like a step above
merely specifying how a job is run.