Add sdkmr.save_and_return_job
Needed for mocking the case where a Job object has the save() method
called on it and is then continued to be used. The save() method has a
side effect of modifying the Job object to add an `.id` field, and so to
mock that the method that calls save() on the job must return the updated job.