Integrate the job requirement resolver
Required a fairly large refactor of the run_job pathway in EE2RunJob
and Condor. CatalogUtils is no longer necessary as a reesult.

Also corrected the spec - in multiple places in the code memory is
treated as MB and disk as GB, and that's how it's defined in the deploy
file.

Quite a bit of argument checking has been moved prior to making any DB
changes - in particular several checks are run prior to creating
the parent job in run_batch.

Condor.run_job now simply translates a job spec into a submission dict.

ee2_scheduler_test has been made redundant by Condor_test, but is left
in to show the tests pass and the changes to said tests to get them to
pass. It will be deleted in the next commit.
for htcondor, making it quite a bit simpler.