Make app_id optional and relax requirements
There are some KBase processes (KBParallel and Narrative downloads,
perhaps more) where it doesn't make sense to provide an app_id but app_id is
currently a required field. This PR:
- makes app_id an optional field
- allows a . rather than a / as the app ID separator as Narrative downloads use
  the method ID as the app ID
- allows just the module name as the app ID as KBParallel sends that

Based on an analysis of the production EE2 DB, there aren't any other regular
cases of bad app IDs.

In the future, as processes stop submitting incorrect app_ids, we may be able to
tighten the restrictions on app IDs again.