Don't toggle mongo collection instance variable
The current MongoUtil implementation has one `mongo_collection` instance
variable that can be set to either the jobs collection or the logs collection
depending on which method was called last. Methods that don't explicitly
choose the collection themselves will use the collection set by whichever
method was most recently called, which might be the wrong collection.

This commit makes two separate variables and ensures the right
collection is being used for each method. It also deletes a number of
unused methods.